Ensure that type names used in extensions are correctly parsed.
We use the internal runtime function `_typeName(_:qualified:)` to get fully
qualified type names in a few places. If a type is referenced from an extension
to that type in another module, the resulting string is prefixed with something
like `"(extension in OtherModule):"` which messes with test plan graphing and
our general expectations about the names of types.

This PR normalizes how we get fully-qualified type names so that we always go
through a particular function that is aware of this prefix and knows how to
remove it.