Allow filtering tests by arbitrary conditions. (#231)
This PR replaces the `Configuration.TestFilter` function typealias with
a nominal type (a structure) that allows us to apply arbitrary predicate
functions to tests.

A simple predicate function is insufficient because it does not track
enough information about a test and its position in a test graph. For
instance, to only run tests tagged `"hello"`, we might naïvely write a
test filter function like:

```swift
configuration.testFilter = { test in
	test.tags.contains("hello")
}
```

But this will result in the test's _containing_ suite (or, recursively,
suites) to not be included in the test plan. The test function will run,
but events will not be generated for its enclosing scopes. This makes
such functions appear to be global/free functions, which they are not.
Badness ensues if you're writing a test harness or CI plugin that
expects events for containing types to be delivered before and after
member test functions.

This is generally solved when filtering by known test ID because we have
an internal structure, `Test.ID.Selection`, that we use to fuzzy-match
test IDs. However, that type requires us to know the set of matching
test IDs upfront, whereas filtering by tag cannot be performed until
_after_ we know the set of tests we might want to run.

The new type contains, in effect, two internal abstractions: one that
takes a test ID selection and uses it to trivially filter a test graph,
and one that takes a function, applies it to each test, and uses it to
construct a test ID selection (that it then uses to filter the test
graph). There's probably a more algorithmically efficient way to do
this, but I've been unable to come up with something that still produces
the expected results for complex test graphs and filtering conditions.

Resolves rdar://119634169.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.