Allow combining `TestFilter` instances with `OR`-like semantics.
This PR extends the `combining(with:)` and `combine(with:)` functions of
`Configuration.TestFilter` to allow specifying whether combinations use `AND`-
or `OR`-like semantics.

Because of how test graphs are constructed, these operations aren't trivially
equivalent to the Swift standard library `&&` and `||` operators defined for
boolean operands, but they have the same overall effect.

This PR introduces a new enumeration in `Configuration.TestFilter` that covers
the two operators; I had at one point in development declared public/SPI
operator functions so that you could write `f1.combine(with: f2, using: &&)`,
but the operator functions had the signature `(Test?, Test?) -> Test?` which
while correct for our use cases would probably _not_ be correct in the general
context of comparing two instances of `Test?` and could be accidentally misused.
An enumeration, while slightly less _Swifty!™_, is unambiguous.

We do not currently make use of `OR`-like semantics anywhere in the testing
library, but adding them opens up new possibilities for tools like Swift Package
Manager in the future.