Replace unsafe '-define-availability' flags with experimental AvailabilityMacro feature (#246)
This replaces the `-define-availability` flags in this package's
`Package.swift` with the pseudo-experimental `AvailabilityMacro=...`
feature.

### Motivation:

We want to use availability macros throughout swift-testing's codebase
to simplify things and reduce the possibility of errors or
inconsistencies. So far, this has only been possible by passing the
`-define-availability` frontend compiler flag, but passing flags this
way is considered unsafe and cannot be used in release versions of the
package. This means that each time we prepare a new tag for release, we
have needed to perform a manual step of replacing each availability
macro with its actual versions, and we'd prefer to not need to do this
step.

### Modifications:

- Replace all `-define-availability` settings with
`AvailabilityMacro=...` experimental feature.
- Remove a now-obsolete step from our release preparation procedure
documentation.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.