Add experimental support for event streaming using JSON.
This PR adds a new command-line argument to the SwiftPM entry point that allows
callers to stream snapshots of events to a file or named pipe. Note that this
argument has not (yet) been added to `swift test`, so to use it you must first
call `swift build --build-tests`, then invoke the executable build product
directly.

This functionality would be useful for clients like VSCode that want to gather
detailed, structured output as it is generated. The existing XML-based output is
of course structured, but is only available and parseable after the test process
finishes, and may fail to decode if the test process crashes.

The output format consists of JSON blobs representing each event along with its
associated test and test case (where applicable.) The events are encoded to
JSON, then the size of the JSON is transmitted as a 64-bit little-endian integer
followed by the JSON itself. The lengths are binary and UTF-8 is assumed for
the JSON. (We considered using [JSONL](https://jsonlines.org/), but it's
significantly harder to decode when using e.g. standard C file streams, which
makes it harder for us to test.)

Any writeable file path can be passed. A regular file can be used, however most
platforms will not allow reading from a regular file while it is open for
writing. Instead, on platforms where it is supported, a named pipe can be used.
On Darwin and Linux, you can use [`mkfifo()`](https://pubs.opengroup.org/onlinepubs/9699919799/functions/mkfifo.html).
On Windows, you can use [`CreateNamedPipeA()`](https://learn.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-createnamedpipea).
Bash and Windows PowerShell have equivalents that can be used instead of C
functions if needed.