Introduce a snapshot type for Error values (#264)
This introduces a new serializable "snapshot" type for representing
`Error` values.

### Motivation:

The `Issue.Kind.errorCaught(_ error: any Error)` enum case stores an
`Error` value, but when taking a snapshot of an `Issue`, we currently
only capture a textual description of the error because `Error` is not
`Codable`. Sometimes, though, is it is useful to have a real `Error`
value on the receiving (i.e. deserializing) end of the snapshot, even if
it is not identical to the original error, for the purpose of
propagating that error or representing it via mechanisms which expect an
`Error` value.

### Modifications:

- Add `ErrorSnapshot` as SPI.
- Replace the `String` associated value describing an error in
`Issue.Kind.Snapshot.errorCaught(_:)` with `ErrorSnapshot`.
- Add a derived `error` property to `Issue.Snapshot`, returning the
`ErrorSnapshot` (if any).

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.