Move `Test.Clock` to SPI groups `Experimental` and `ForToolsIntegrationOnly`.
This PR moves `Test.Clock` to the `Experimental` and `ForToolsIntegrationOnly`
SPI groups per the SPI policy outlined
[here](https://github.com/apple/swift-testing/blob/main/Documentation/SPI.md).

`Test.Clock` is a candidate for promotion to API, so it is available both for
tools integration and as an experimental interface for developer evaluation.

This PR also moves the `Expectation.mismatchedErrorDescription` and
`.differenceDescription` properties to the `ForToolsIntegrationOnly` SPI group
as they are the only remaining symbols in the `ExperimentalEventHandling` group.