Diagnose ambiguous uses of `#require()` with an argument of type `Bool?`.
This PR adds new diagnostics when `#require()` is passed an argument of type
`Bool?` because such an argument may be ambiguously interpreted. If the test
author writes:

```swift
try #require(love?.isFleeting)
```

Do they mean to unwrap the optional boolean value and return it, or do they mean
to test if `isFleeting` is true or false?

With this change, an expression such as that one that returns a value of type
`Bool?` will produce a diagnostic of the form:

<SCRN>

Resolves rdar://105727852.