Diagnose ambiguous uses of `#require()` with an argument of type `Bool?`. (#276)
This PR adds new diagnostics when `#require()` is passed an argument of
type `Bool?` because such an argument may be ambiguously interpreted. If
the test author writes:

```swift
try #require(love?.isFleeting)
```

Do they mean to unwrap the optional boolean value and return it, or do
they mean to test if `isFleeting` is true or false?

With this change, an expression such as that one that returns a value of
type `Bool?` will produce a diagnostic of the form:

![Screenshot 2024-03-08 at 4 41
30 PM](https://github.com/apple/swift-testing/assets/4145863/248105f8-1d48-4c9b-96b9-e6e3b066b4ca)

(The diagnostics are presented as fix-its in Xcode.)

Resolves rdar://105727852.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.