Adopt `lexicalContext` from swift-syntax-6.0.0.
This PR conditionally adopts the new `lexicalContext` member of
`MacroExpansionContext` added in https://github.com/apple/swift-syntax/pull/1554.
If the SwiftSyntax600 pseudo-module is available, then this member should be
available for use and can be used to perform additional diagnostics for tests
and to get the names of their containing types.

With this PR, if built against an older toolchain (5.11 or earlier), the old
hacky "is there leading whitespace?" mechanism is still used.

A future PR will recursively perform suite-level diagnostics on the lexical
contexts containing tests and suites, so that a test cannot be (easily) inserted
into a type that cannot be used as a suite.

Resolves rdar://109439578.