Make event recorder types non-copyable.
This PR makes the event recorder types in swift-testing non-copyable. They can
still be captured by an event handler without issue, but cannot be moved around.
An event recorder carries with it a large amount of data (the statistics around
recorded events that are used to inform later events) and copying them, even
inadvertently, can be very expensive.