Add a `comment` property to the `.bug()` trait.
This PR adds an optional `comment` property to the `.bug()` trait so that
developers can associate additional notes that may not be obvious from just a
bug's identifier. Since the testing library can't go off crawling arbitrary bug
trackers looking for metadata, it makes sense to include basic metadata inline.
For example, since this PR is resolving rdar://124101140 whose description in
Apple's Radar system is "Add an optional Comment parameter to `.bug` trait", an
associated test could have this trait added to it:

```swift
.bug("rdar://124101140", "Add an optional Comment parameter to `.bug` trait")
```

This PR also adds `Codable` conformance to `Bug` so that it can be encoded and
decoded when needed.