Drop remaining 5.9 support. (#288)
This PR updates the package manifest to require Swift 5.10 or newer. We
previously had a tools version of 5.9 so that content from the package
would be indexed by Swift Package Index, but they have updated their
systems to use 5.10 so this is no longer required.

As a reminder, Swift 5.10 support is temporary for the benefit of
developers who cannot rely on a nightly main-branch toolchain. **Swift
5.10 support will be removed in a future update.**

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.