Adopt `TypeInfo` in `Test`. (#310)
This PR adopts `TypeInfo` in `Test` so that we're not carrying around a
direct type reference. This will allow us to synthesize instances of
`Test` at runtime that describe types for which we don't have the
metatype pointer. It may also enable future Embedded Swift work (where
`Any.Type` is unavailable.)

This change updates `TypeInfo` to track either a metatype (where
available) _or_ a qualified/unqualified name pair. If you start with a
metatype, you keep it around until you need to encode the instance, at
which point we encode strings, and subsequent decoding gives you back a
`TypeInfo` that's string-backed. This should improve memory usage and
storage requirements because otherwise every instance of `Test` in
memory will need to keep strings around and they are unlikely to be
interned/uniqued in the process.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.