Flatten call stack when using `TypeInfo`. (#318)
This PR removes the internal `nameComponents(of:)`,
`fullyQualifiedName(of:)`, `isSwiftEnumeration(_:)`, and
`isImportedFromC(_:)` functions and moves their implementations into the
appropriate spots inside of `TypeInfo`. This change ensures that our
code always uses `TypeInfo` to get the names of types.

This PR also implements a _de facto_ implementation for
`String.init(describingForTest:)` for metatypes that calls through to
the same code. In the future, we might want to allow an additional
`verbose` argument to that initializer that causes `fullyQualifiedName`
to be produced instead of `unqualifiedName`.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.