Add a dedicated `Win32Error` type for errors in the Windows domain.
On Windows, errors produced by Win32 functions (`GetLastError()`) are in a
separate domain from errors produced by C standard library functions (`errno`).

Although we're not currently making _much_ use of Win32 API that produces errors
in this domain, we do use it in `Environment` and #307 uses it, so I've factored
out the part of that PR that adds a dedicated error type.