Use `lexicalContext` to improve diagnostics for tag declarations.
This PR uses the swift-syntax-600 `lexicalContext` property to enforce the
compile-time requirement that tags declared with `@Tag` must be declared in an
extension to `Tag` or in a type declared inside `Tag`. Fix-its are provided for
a few scenarios. The new diagnostics are:

```swift
@Tag var x: Tag // 🛑 Attribute 'Tag' cannot be applied to a global variable
                // Declare in an extension to 'Tag' [Fix]
                // Remove attribute 'Tag' [Fix]

extension String {
  @Tag var x: Tag // 🛑 Attribute 'Tag' cannot be applied to a property except in an extension to 'Tag'
}

extension Tag {
  @Tag var x: Tag // 🛑 Attribute 'Tag' cannot be applied to an instance property
                  // Add 'static' [Fix]
}

extension Tag {
  @Tag var x: String // 🛑 Attribute 'Tag' cannot be applied to a property of type 'String'
                     // Change type to 'Tag' [Fix]
                     // Remove attribute 'Tag' [Fix]
}
```

As with #332 and #333, these problems are only diagnosed when using
swift-syntax-600.