Improve handling of required expectation failures in error matcher closure of `#expect { } throws: { }` (#305)
This improves the handling of `#require` failures in the `throws: {
error in … }` error-matching closure of an `#expect { } throws: { }`
expectation, by ensuring the resulting issues are relevant.

Take this example:

```swift
struct AnotherError: Error {}
struct MyError: Error { ... }

@Test func example() {
  #expect {
    throw AnotherError()
  } throws: { error in
    let myError = try #require(error as? MyError)
    return myError.someProperty
  }
}
```

Before this PR, the following issues were recorded:
1. ❌ Expectation failed: (error → AnotherError()) as? (MyError →
AnotherError)
2. ❌ Caught error: ExpectationFailedError(expectation:
Testing.Expectation(...))
3. ❌ Expectation failed: a second error
"ExpectationFailedError(expectation: Testing.Expectation(...))" was
thrown when checking error "AnotherError()"

Issue 1 is from the `#require`, and it's correct and appropriate. But
issue 2 is unnecessary noise; it's revealing an implementation detail
about the underlying error which all `#require` expectations throw when
they fail (`ExpectationFailedError`), so this issue ought to be
suppressed. Issue 3 is _meant_ to indicate the failure of the overall
`#expect { } throws: { }`, but it's phrased in a misleading way which
focuses on the internal error mentioned earlier, instead of describing
the real problem: that the caught error did not match.

After this PR, the following issues are recorded instead:
- ❌ Expectation failed: (error → AnotherError()) as? (MyError →
AnotherError)
- ❌ Expectation failed: unexpected error "AnotherError()" was thrown

Along the way, I consolidated some logic into a more widely-shared
helper which invokes closures which may throw and translates caught
errors into issues.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.