Allow non-final classes as test suites.
With our previous dependency on swift-syntax-510, we were unable to tell the
precise type name of a type containing a test function, so we substituted `Self`
in place of the type name. This mostly worked, except when the type was a
non-final class (in which case `Self` was covariant and would cause errors at
compile time.)

Now that we're using swift-syntax-600 and `lexicalContext`, we no longer have
this issue, so we can lift the requirement that classes be `final`.

Note that, unlike XCTest, tests in a class in swift-testing are not inherited by
subclasses thereof. XCTest relies on the Objective-C runtime (on Darwin) or
compile-time analysis of the build product (when using corelibs-xctest) to do
test discovery, whereas swift-testing relies on the Swift runtime. Inherited
tests/suites are a future direction.