Remove the new diagnostic added in #302.
In #302, we added a compile-time diagnostic (warning) for expressions like:

```swift
```

Because we figured that the lack of expression expansion for effectful
expressions might be confusing. However, we've found that the diagnostic is
significantly noisier than we'd like and the cons outweigh the pros. Hence, this
PR removes that diagnostic.

Resolves rdar://126393932.