Remove the new diagnostic added in #302. (#352)
In #302, we added a compile-time diagnostic (warning) for expressions
like:

```swift
#expect(try await foo())
```

Because we figured that the lack of expression expansion for effectful
expressions might be confusing. However, we've found that the diagnostic
is significantly noisier than we'd like and the cons outweigh the pros.
Hence, this PR removes that diagnostic.

Resolves rdar://126393932.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.