Refactor `__swiftPMEntryPoint()` so that it can be used with arbitrary package versions.
This PR refactors `__swiftPMEntryPoint()` such that it takes an instance of a
new type, `__CommandLineArguments_v0`. This type conforms to `Codable` and
represents the parsed command-line arguments that the package can understand.
Normally, swift-testing will construct an instance of this type from the
command-line arguments to the current process, but a caller can pass in a
different instance instead.

> [!WARNING]
> The structure of `__CommandLineArguments_v0`, how it encodes to/from JSON, and
> all other details about it are still subject to change. The hope is that it
> becomes a stable ABI surface _eventually_, but we're not there yet. :)

This PR also introduces a function, also intended as part of our stable ABI,
`swt_copyABIEntryPoint_v0()`, that can be looked up at runtime and used to hook
into the testing library when a test target is loaded into the current process.
It takes a JSON blob (encoded from `__CommandLineArguments_v0`) as input and
yields events/event contexts in the form of two JSON blobs that can be decoded
in a callback (the structure of these blobs is _also_ not finalized, but would
be part of our stable ABI eventually too.)

Because trying to load a Swift function dynamically at runtime is Hard™,
`swt_copyABIEntryPoint_v0()` is a C function that yields the "real" Swift
function. This indirection also gives us the flexibility to produce a thick
Swift closure rather than a thin function.